Consolidate ruby to rust array conversions
